Pass friendly name to chromecast-scanner
Chromecasts no longer appear to contain the friendly name in their mDNS
hostnames. Instead, the friendly name appears only in each device's TXT
record.

- Remove the .local suffix when passing a specific device name to
  chromecast-scanner since we can no longer hope to guess the device's
  hostname base on its friendly name.

Passing the raw device name to the scanner project allows the scanner
project to search for devices with matching friendly names regardless of
any future friendly name format/advertising changes.

Note that this requires a supporting change from chromecast-scanner
dependency to search the TXT records for the friendly name field. However,
the .local suffixing is not currently working at all, so this is not a
breaking change.